Moved model definition classes to models.py
